+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.9.0-beta.2 (Unreleased)
+## 4.10.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -13,6 +13,12 @@
 ### Other Changes
 
 - The default service version is now `2026-03-01-preview`.
+
+## 4.9.0 (2026-05-26)
+
+### Features Added
+
+- Includes all changes from 4.9.0-beta.1.
 
 ## 4.9.0-beta.1 (2026-04-01)